+++ pkg/ddl/column.go
@@ -104,100 +104,6 @@ func checkAddColumn(t *meta.Meta, job *model.Job) (*model.TableInfo, *model.Colu
 	return tblInfo, columnInfo, col, pos, false, nil
 }
 
-func onAddColumn(d *ddlCtx, t *meta.Meta, job *model.Job) (ver int64, err error) {
-	// Handle the rolling back job.
-	if job.IsRollingback() {
-		ver, err = onDropColumn(d, t, job)
-		if err != nil {
-			return ver, errors.Trace(err)
-		}
-		return ver, nil
-	}
-
-	failpoint.Inject("errorBeforeDecodeArgs", func(val failpoint.Value) {
-		//nolint:forcetypeassert
-		if val.(bool) {
-			failpoint.Return(ver, errors.New("occur an error before decode args"))
-		}
-	})
-
-	tblInfo, columnInfo, colFromArgs, pos, ifNotExists, err := checkAddColumn(t, job)
-	if err != nil {
-		if ifNotExists && infoschema.ErrColumnExists.Equal(err) {
-			job.Warning = toTError(err)
-			job.State = model.JobStateDone
-			return ver, nil
-		}
-		return ver, errors.Trace(err)
-	}
-	if columnInfo == nil {
-		columnInfo = InitAndAddColumnToTable(tblInfo, colFromArgs)
-		logutil.DDLLogger().Info("run add column job", zap.Stringer("job", job), zap.Reflect("columnInfo", *columnInfo))
-		if err = checkAddColumnTooManyColumns(len(tblInfo.Columns)); err != nil {
-			job.State = model.JobStateCancelled
-			return ver, errors.Trace(err)
-		}
-	}
-
-	originalState := columnInfo.State
-	switch columnInfo.State {
-	case model.StateNone:
-		// none -> delete only
-		columnInfo.State = model.StateDeleteOnly
-		ver, err = updateVersionAndTableInfoWithCheck(d, t, job, tblInfo, originalState != columnInfo.State)
-		if err != nil {
-			return ver, errors.Trace(err)
-		}
-		job.SchemaState = model.StateDeleteOnly
-	case model.StateDeleteOnly:
-		// delete only -> write only
-		columnInfo.State = model.StateWriteOnly
-		ver, err = updateVersionAndTableInfo(d, t, job, tblInfo, originalState != columnInfo.State)
-		if err != nil {
-			return ver, errors.Trace(err)
-		}
-		// Update the job state when all affairs done.
-		job.SchemaState = model.StateWriteOnly
-	case model.StateWriteOnly:
-		// write only -> reorganization
-		columnInfo.State = model.StateWriteReorganization
-		ver, err = updateVersionAndTableInfo(d, t, job, tblInfo, originalState != columnInfo.State)
-		if err != nil {
-			return ver, errors.Trace(err)
-		}
-		// Update the job state when all affairs done.
-		job.SchemaState = model.StateWriteReorganization
-		job.MarkNonRevertible()
-	case model.StateWriteReorganization:
-		// reorganization -> public
-		// Adjust table column offset.
-		failpoint.InjectCall("onAddColumnStateWriteReorg")
-		offset, err := LocateOffsetToMove(columnInfo.Offset, pos, tblInfo)
-		if err != nil {
-			return ver, errors.Trace(err)
-		}
-		tblInfo.MoveColumnInfo(columnInfo.Offset, offset)
-		columnInfo.State = model.StatePublic
-		ver, err = updateVersionAndTableInfo(d, t, job, tblInfo, originalState != columnInfo.State)
-		if err != nil {
-			return ver, errors.Trace(err)
-		}
-
-		// Finish this job.
-		job.FinishTableJob(model.JobStateDone, model.StatePublic, ver, tblInfo)
-		addColumnEvent := statsutil.NewAddColumnEvent(
-			job.SchemaID,
-			tblInfo,
-			[]*model.ColumnInfo{columnInfo},
-		)
-		asyncNotifyEvent(d, addColumnEvent)
-	default:
-		err = dbterror.ErrInvalidDDLState.GenWithStackByArgs("column", columnInfo.State)
-	}
-
-	return ver, errors.Trace(err)
-}
-
 // CheckAfterPositionExists makes sure the column specified in AFTER clause is exists.
 // For example, ALTER TABLE t ADD COLUMN c3 INT AFTER c1.
 func CheckAfterPositionExists(tblInfo *model.TableInfo, pos *ast.ColumnPosition) error {
